users/profile-service/app/__init__.py
from flask import Flask
from .db import db
from .routes import profile_bp
from dotenv import load_dotenv
import os

def create_app():
    load_dotenv()
    app = Flask(__name__)
    app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv("DATABASE_URL")
    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False

    db.init_app(app)
    app.register_blueprint(profile_bp)

    with app.app_context():
        db.create_all()

    return app
